pyproject.toml
[project]
name = "video-transcriber"
version = "0.1.0"
description = "Generates sub-title for the input video"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "assemblyai>=0.48.2",
    "celery>=5.6.0",
    "django==5.0",
    "django-enum>=2.3.0",
    "djangorestframework>=3.16.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular[sidecar]>=0.29.0",
    "ffmpeg-python>=0.2.0",
    "openai>=2.9.0",
    "python-magic-bin>=0.4.14",
    "redis>=7.1.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.8",
]

[tool.ruff]
line-length = 125
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
testpaths = ["tests"]
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]
pythonpath = ["src"]
